Fix(Issue Tracker): Abstract common issue creation functionality
