+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/JiraServerChannel.java
@@ -29,7 +29,6 @@ import org.springframework.stereotype.Component;
 
 import com.google.gson.Gson;
 import com.synopsys.integration.alert.channel.jira.common.JiraMessageContentConverter;
-import com.synopsys.integration.alert.channel.jira.server.JiraServerRequestDelegator;
 import com.synopsys.integration.alert.common.channel.IssueTrackerChannel;
 import com.synopsys.integration.alert.common.channel.issuetracker.config.IssueTrackerContext;
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerRequest;
@@ -59,7 +58,7 @@ public class JiraServerChannel extends IssueTrackerChannel {
 
     @Override
     protected List<IssueTrackerRequest> createRequests(IssueTrackerContext context, DistributionEvent event) throws IntegrationException {
-        return jiraContentConverter.convertMessageContents(context.getIssueConfig(), event.getContent());
+        return jiraContentConverter.createRequests(context.getIssueConfig(), event.getContent());
     }
 
     @Override
